Started rewrite, bah
Yeah, apparently the old PulseDelay wasn't quite as rock-solid as I recall.
So I'm going to rewrite the main processing algorithm and try to be a bit
more clever in the process. Here the initial architecture and plugin state
machine is implemented, but switching between plugin states doesn't quite
work as expected.